Finalized new pickupable sync component and removed old net ID system

+++ src/MSCMPMessages/Messages/PickupableActivateMessage.cs
@@ -1,7 +1,7 @@
 ﻿namespace MSCMPMessages.Messages {
 	[NetMessageDesc(MessageIds.PickupableActivate)]
 	class PickupableActivateMessage {
-		ushort id;
+		int id;
 		bool activate;
 	}
 }
