+++ src/MSCMPMessages/Messages/PickupableDestroyMessage.cs
@@ -1,6 +1,6 @@
 namespace MSCMPMessages.Messages {
 	[NetMessageDesc(MessageIds.PickupableDestroy)]
 	class PickupableDestroyMessage {
-		ushort id;
+		int id;
 	}
 }
