+++ src/MSCMPMessages/Messages/PickupableSetPositionMessage.cs
@@ -1,7 +1,7 @@
 ﻿namespace MSCMPMessages.Messages {
 	[NetMessageDesc(MessageIds.PickupableSetPosition)]
 	class PickupableSetPositionMessage {
-		ushort id;
+		int id;
 		Vector3Message position;
 	}
 }
